requirements.txt
bidict==0.22.0
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
et-xmlfile==1.1.0
Flask==2.1.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
fonttools==4.37.4
greenlet==1.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.0.221010
Pillow==9.2.0
PyAstronomy==0.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
quantities==0.13.0
requests==2.28.1
scipy==1.9.2
six==1.16.0
tqdm==4.64.1
types-pytz==2022.4.0.0
typing_extensions==4.4.0
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.2.1
